Added authorization functionality

+++ application/messages/views.py
@@ -1,6 +1,6 @@
-from application import app, db
+from application import app, db, login_required, login_manager
 from flask import render_template, request, redirect, url_for
-from flask_login import login_required, current_user
+from flask_login import current_user
 from datetime import datetime
 
 from application.messages.models import Message
@@ -17,26 +17,30 @@ def message_details(message_id):
     message = Message.query.get(message_id)
     return render_template("messageDetails.html", message=message)
 
-# poistaa yksittäisen viestin
+# poistaa yksittäisen viestin, sallittu vain viestin kirjoittaneelle
 @app.route("/messages/delete/<message_id>", methods=["POST"])
-@login_required
+@login_required(role="BASIC")
 def messages_delete(message_id):
     message_to_be_deleted = Message.query.get(message_id)
+    if message_to_be_deleted.author_id != current_user.id:
+        return login_manager.unauthorized()
     db.session.delete(message_to_be_deleted)
     db.session.commit()
     return redirect(url_for("index"))
 
 # palauttaa lomakkeen viestin editoimiseen (väliaikainen ratkaisu)
 @app.route("/messages/edit/<message_id>", methods=["GET"])
-@login_required
+@login_required(role="BASIC")
 def messages_editingForm(message_id):
     return render_template("messageEditingForm.html", form=MessageForm(), message_id=message_id)
 
-# korvaa viestin uudella (editointitoiminnallisuus)
+# korvaa viestin uudella (editointitoiminnallisuus), sallittu vain viestin kirjoittaneelle
 @app.route("/messages/edit/<message_id>", methods=["POST"])
-@login_required
+@login_required(role="BASIC")
 def messages_edit(message_id):
     message = Message.query.get(message_id)
+    if message.author_id != current_user.id:
+        return login_manager.unauthorized()
     form = MessageForm(request.form)
     if not form.validate():
         return render_template("messageEditingForm.html", form=MessageForm(), message_id=message_id)
